+++ caller/vcdl-caller/nativeMessaging.h
@@ -3,7 +3,7 @@
 #define     MAX_MSG_LEN     (1024*1024)
 
 // send msg to extension
-size_t sendToExtension(const char* msg);
+uint32_t sendToExtension(const char* msg);
 
 // recieve msg from extension
-char* recieveFromExtension(size_t* len);
+char* recieveFromExtension(uint32_t* len);
